infra: tip_selection folder
