+++ consensus/BUILD
@@ -29,10 +29,6 @@ cc_library(
     deps = [
         "//common:errors",
         "//consensus/bundle_validator",
-        "//consensus/cw_rating_calculator",
-        "//consensus/entry_point_selector",
-        "//consensus/exit_probability_randomizer",
-        "//consensus/exit_probability_validator",
         "//consensus/ledger_validator",
         "//consensus/milestone:milestone_service",
         "//consensus/milestone:milestone_tracker",
@@ -41,7 +37,11 @@ cc_library(
         "//consensus/snapshot:snapshots_service",
         "//consensus/snapshot/local_snapshots:local_snapshots_manager",
         "//consensus/tangle",
-        "//consensus/tip_selector",
+        "//consensus/tip_selection:tip_selector",
+        "//consensus/tip_selection/cw_rating_calculator",
+        "//consensus/tip_selection/entry_point_selector",
+        "//consensus/tip_selection/exit_probability_randomizer",
+        "//consensus/tip_selection/exit_probability_validator",
         "//consensus/transaction_solidifier",
         "//consensus/transaction_validator",
     ],
