+++ consensus/consensus.h
@@ -11,16 +11,16 @@
 #include "common/errors.h"
 #include "consensus/bundle_validator/bundle_validator.h"
 #include "consensus/conf.h"
-#include "consensus/cw_rating_calculator/cw_rating_calculator.h"
-#include "consensus/entry_point_selector/entry_point_selector.h"
-#include "consensus/exit_probability_randomizer/exit_probability_randomizer.h"
-#include "consensus/exit_probability_validator/exit_probability_validator.h"
 #include "consensus/ledger_validator/ledger_validator.h"
 #include "consensus/milestone/milestone_service.h"
 #include "consensus/milestone/milestone_tracker.h"
 #include "consensus/snapshot/local_snapshots/local_snapshots_manager.h"
 #include "consensus/snapshot/snapshot.h"
-#include "consensus/tip_selector/tip_selector.h"
+#include "consensus/tip_selection/cw_rating_calculator/cw_rating_calculator.h"
+#include "consensus/tip_selection/entry_point_selector/entry_point_selector.h"
+#include "consensus/tip_selection/exit_probability_randomizer/exit_probability_randomizer.h"
+#include "consensus/tip_selection/exit_probability_validator/exit_probability_validator.h"
+#include "consensus/tip_selection/tip_selector.h"
 #include "consensus/transaction_solidifier/transaction_solidifier.h"
 #include "consensus/transaction_validator/transaction_validator.h"
