--- consensus/tip_selector/BUILD
@@ -1,19 +1,0 @@
-cc_library(
-    name = "tip_selector",
-    srcs = ["tip_selector.c"],
-    hdrs = ["tip_selector.h"],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//common:errors",
-        "//common/trinary:trit_array",
-        "//consensus:model",
-        "//consensus/cw_rating_calculator",
-        "//consensus/entry_point_selector",
-        "//consensus/exit_probability_randomizer",
-        "//consensus/exit_probability_validator",
-        "//consensus/ledger_validator",
-        "//consensus/milestone:milestone_tracker",
-        "//consensus/tangle",
-        "//utils:logger_helper",
-    ],
-)
